returned deploy.ts to how it was previously and made requested changes to deploy-contract.ts
